Reshuffle code to include signal in ped and coh noise
